+++ index.js
@@ -408,6 +408,7 @@ class Jetpack {
   async package() {
     const { service } = this.serverless;
     const servicePackage = service.package;
+    const serviceIsNode = (service.provider.runtime || "").startsWith("node");
     const { concurrency } = this._serviceOptions;
     const options = this.options || {};
     const report = !!options.report;
@@ -435,7 +436,9 @@ class Jetpack {
       }))
       .map((obj) => ({
         ...obj,
-        functionPackage: obj.functionObject.package || {}
+        functionPackage: obj.functionObject.package || {},
+        runtime: obj.functionObject.runtime,
+        isNode: (obj.functionObject.runtime || "").startsWith("node")
       }))
       .map((obj) => ({
         ...obj,
@@ -446,7 +449,12 @@ class Jetpack {
 
     // Get list of individual functions to package.
     const fnsPkgsToPackage = fnsPkgs.filter((obj) =>
-      (servicePackage.individually || obj.individually) && !(obj.disable || obj.artifact)
+      // Individually packaged
+      (servicePackage.individually || obj.individually)
+      // Enabled
+      && !(obj.disable || obj.artifact)
+      // Function runtime is node or unspecified + service-level node.
+      && (obj.isNode || !obj.runtime && serviceIsNode)
     );
     const numFns = fnsPkgsToPackage.length;
     tasks = tasks.concat(fnsPkgsToPackage.map((obj) => () =>
@@ -457,6 +465,8 @@ class Jetpack {
     // to package the service or not.
     const shouldPackageService = !servicePackage.individually
       && !servicePackage.artifact
+      // Service must be Node.js
+      && serviceIsNode
       // Don't package service if we specify a single function **and** have a match
       && (!singleFunctionName || !numFns)
       // Otherwise, have some functions left that need to use the service package.